services/image-service/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn==0.34.0

# Database
motor==3.7.0
pymongo==4.11.3

# Auth
PyJWT==2.10.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1

# HTTP client (for LLM service calls)
httpx==0.28.1

# Image processing
Pillow==11.1.0
python-magic==0.4.27
python-multipart==0.0.20
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
mongomock-motor==0.0.34
